requirements.txt
asgiref==3.8.1
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
drf-yasg==1.21.7
fonttools==4.52.4
inflection==0.5.1
joblib==1.4.2
jsonfield==3.1.0
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
